+++ pacman/model/graphs/machine/machine_vertex.py
@@ -61,6 +61,25 @@ class MachineVertex(AbstractVertex):
         :rtype: ~pacman.model.resources.ResourceContainer
         """
     @property
-    @overrides(AbstractVertex.timestep_in_us)
     def timestep_in_us(self):
         return self._timestep_in_us
+
+    def simtime_in_us_to_timesteps(self, simtime_in_us):
+        """
+        Helper function to convert simtime in us to whole timestep
+
+        This function verfies that the simtime is a multile of the timestep.
+
+        :param simtime_in_us: a simulation time in us
+        :type simtime_in_us: int
+        :return: the exact number of timeteps covered by this simtime
+        :rtype: int
+        :raises ValueError: If the simtime is not a mutlple of the timestep
+        """
+        n_timesteps = simtime_in_us // self.timestep_in_us
+        check = n_timesteps * self.timestep_in_us
+        if check != simtime_in_us:
+            raise ValueError(
+                "The requested time {} is not a multiple of the timestep {}"
+                "".format(simtime_in_us, self.timestep_in_us))
+        return n_timesteps
