+++ uinit_test_objects/simple_test_vertex.py
@@ -16,7 +16,6 @@
 """ test vertex used in many unit tests
 """
 from spinn_utilities.overrides import overrides
-from pacman.model.graphs import AbstractVertex
 from pacman.model.graphs.application import ApplicationVertex
 from pacman.model.graphs.machine import SimpleMachineVertex
 from pacman.model.resources import (
@@ -93,6 +92,6 @@ class SimpleTestVertex(ApplicationVertex):
         return self._n_atoms
 
     @property
-    @overrides(AbstractVertex.timestep_in_us)
-    def timestep_in_us(self):
-        return 1000
+    @overrides(ApplicationVertex.timesteps_in_us)
+    def timesteps_in_us(self):
+        return set()
